+++ src/backend/ta/motion-tracking/test-img-output/bert_output.png
