+++ src/backend/ta/motion-tracking/test-img-output/bodrum_output.png
