+++ src/backend/ta/motion-tracking/test-img/ask3.jpg
