+++ src/backend/ta/motion-tracking/test-img/ask_2.jpg
