+++ src/backend/ta/motion-tracking/test-img/ask_4.jpg
