+++ src/backend/ta/motion-tracking/test-img/askpalm.jpg
