+++ src/backend/ta/motion-tracking/test-img/hand_test.jpg
